functions/requirements.txt
google-cloud-bigquery
Flask==1.0.2